Fixed the updating of product availability
I had an issue when updating a user's product availability. Users were able to change their product from "In Stock" to "Out of Stock", but couldn't change their product from "Out of Stock" back to "In Stock".

I spent hours trying to figure out whether the integer being passed to the controller was being lost, or the browser was caching the previous availability, or if the <option> list was not being populated correctly, etc.

The issue was in the <select name=""> I set the name to the attribute of the product, for example: <select name="@product.ProdAvailable"> which is not a parameter of the action I was posting the form to.